fix: remove AI from my skills
I think using ChatGPT is useful for learning concepts *supplementarily*
as a tutor, which is what I use it for. However, adding that to my
repetoire of skills on my portfolio seems misleading to me just because
I've used midjourney or ChatGPT a tiny bit in my projects or for
brainstorming help, especially since using ChatGPT for code has proven
to be a fruitless endeavor. From this point forward, I'll make sure to
only add things as skills that I've proven my understanding of
through a finished project, which will be more honest to both myself
and prospective clients. So, once I do a project with AI, I can put
that, and if it specifically uses the ChatGPT API then their logo will
go along with the other "tools" section logos.